bookmark_collector/pyproject.toml
[tool.poetry]
name = "bookmark_collector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
gunicorn = "20.0.4"
pytest = "^7.4.3"
flake8 = "^6.1.0"
bs4 = "^0.0.1"
requests = "^2.31.0"
djoser = "^2.2.2"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
